change: SPM supported
